gateway/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
cryptography==42.0.8
httpx==0.27.0
pydantic-settings==2.3.0
